Bo--工作流国际化和xml修改为下载，图片modal模态框查看，批量删除功能

Bo--工作流国际化和xml修改为下载，图片modal模态框查看，批量删除功能

See merge request !67

+++ eweb/src/main/java/com/dsdl/eidea/sys/web/controller/WorkflowModelController.java
@@ -172,19 +172,27 @@ public class WorkflowModelController {
             }
 
             ByteArrayInputStream in = new ByteArrayInputStream(exportBytes);
-            IOUtils.copy(in, response.getOutputStream());
-
-            response.setHeader("Content-Disposition", "attachment; filename=" + filename);
-            response.flushBuffer();
+            response.reset();
+            response.setContentType("application/x-download");
+            response.addHeader("Content-Disposition", "attachment; filename=" + new String(filename.getBytes("utf-8"),"iso8859-1"));
+            byte[] b = new byte[1024];
+            int len = -1;
+            while ((len = in.read(b, 0, 1024)) != -1) {
+                response.getOutputStream().write(b, 0, len);
+            }
         } catch (Exception e) {
             log.error("导出model的xml文件失败：modelId={}, type={}", modelId, type, e);
         }
     }
-    @RequestMapping(value = "/delete/{modelId}")
+    @RequestMapping(value = "/delete")
     @ResponseBody
     @RequiresPermissions("delete")
-    public JsonResult<List<Model>> delete(@PathVariable("modelId") String modelId) {
-        repositoryService.deleteModel(modelId);
+    public JsonResult<List<Model>> delete(@RequestBody String[] ids) {
+        if(ids != null && ids.length > 0){
+            for(String modelId:ids){
+                repositoryService.deleteModel(modelId);
+            }
+        }
         return modelList();
     }
